[NC-2044] refactor json rpc authentication to be provided as a service (#825)
